requirements.txt
# Core AI and LLM dependencies
autogen==0.2.0
openai>=1.0.0
anthropic>=0.25.0

# Web scraping and search
duckduckgo-search==3.9.3
beautifulsoup4==4.12.2
requests==2.31.0

# Vector database and retrieval
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Web interface
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2

# Data processing
pandas>=1.5.0
numpy>=1.21.0
python-dotenv>=1.0.0

# Image processing (for OCR if needed)
Pillow>=9.0.0
pytesseract>=0.3.10

# Utility packages
urllib3>=1.26.0
lxml>=4.9.0
typing-extensions>=4.0.0 